Fixed info webview url redirects; work on displaying open source notices.
The Info webview now opens all links in the webview itself, instead of
asking Android to open them in a browser.

Added the Google recommended solution for collecting and displaying
license notices of all open source software used in the app. This now
works, but the list is empty. So not fully functional yet.